feat: 회원가입 프로필 선택 페이지 컴포넌트화
